Bluetooth: OTS - rename client files
Rename OTS client files from "otc" to "ots_client".
