+++ subsys/bluetooth/services/ots/ots_client_internal.h
